+++ Arex388.Carmine/Models/Vehicle.cs
@@ -79,10 +79,10 @@ public sealed class Vehicle {
 	public VehicleId Id { get; init; }
 
 	/// <summary>
-	/// The vehicle's last activity timestamp in UTC.
+	/// The vehicle's last activity timestamp.
 	/// </summary>
 	[JsonPropertyName("last_activity")]
-	public DateTime? LastActivityAtUtc { get; init; }
+	public DateTime? LastActivityAt { get; init; }
 
 	/// <summary>
 	/// The vehicle's last trip id.